Gen 1&2 Review Part 2
1. Replace statExperience with ev, changing ev to 16-bit, due to mistaken belief that ev needs to stay 8-bit.
2. @piepie62 improvements for G1 and G2 text handling and adding a nullptr case to getPKM.
3. Extraction of BCD handling to other places for general use (hopefully only to ever be used by Gen I).
4. Clean up handling of trade OT char, and clean up handling of incompatible-language imports.

+++ include/pkx/PB7.hpp
@@ -113,8 +113,8 @@ namespace pksm
         void gender(Gender g) override;
         [[nodiscard]] u16 alternativeForm(void) const override;
         void alternativeForm(u16 v) override;
-        [[nodiscard]] u8 ev(Stat ev) const override;
-        void ev(Stat ev, u8 v) override;
+        [[nodiscard]] u16 ev(Stat ev) const override;
+        void ev(Stat ev, u16 v) override;
         // Stubbed; data no longer exists
         [[nodiscard]] u8 contest(u8) const override { return 0; }
         void contest(u8, u8) override {}
