+++ include/pkx/PK3.hpp
@@ -137,8 +137,8 @@ namespace pksm
         void ability(Ability v) override;
         [[nodiscard]] bool abilityBit(void) const;
         void abilityBit(bool v);
-        [[nodiscard]] u8 ev(Stat ev) const override;
-        void ev(Stat ev, u8 v) override;
+        [[nodiscard]] u16 ev(Stat ev) const override;
+        void ev(Stat ev, u16 v) override;
         [[nodiscard]] u8 contest(u8 contest) const override;
         void contest(u8 contest, u8 v) override;
         [[nodiscard]] bool hasRibbon(Ribbon rib) const override;
