+++ include/pkx/PK6.hpp
@@ -114,8 +114,8 @@ namespace pksm
         void gender(Gender g) override;
         [[nodiscard]] u16 alternativeForm(void) const override;
         void alternativeForm(u16 v) override;
-        [[nodiscard]] u8 ev(Stat ev) const override;
-        void ev(Stat ev, u8 v) override;
+        [[nodiscard]] u16 ev(Stat ev) const override;
+        void ev(Stat ev, u16 v) override;
         [[nodiscard]] u8 contest(u8 contest) const override;
         void contest(u8 contest, u8 v) override;
