+++ include/wcx/WB7.hpp
@@ -117,7 +117,7 @@ namespace pksm
         [[nodiscard]] u8 contest(u8) const override { return 0; }
         [[nodiscard]] u8 awakened(Stat index) const;
         [[nodiscard]] u8 iv(Stat index) const override;
-        [[nodiscard]] u8 ev(Stat index) const;
+        [[nodiscard]] u16 ev(Stat index) const;
         [[nodiscard]] u16 additionalItem(void) const;
         [[nodiscard]] bool hasRibbon(Ribbon) const override { return false; }
         [[nodiscard]] bool ribbon(Ribbon) const override { return false; }
