+++ source/pkx/PK4.cpp
@@ -343,8 +343,8 @@ namespace pksm
     Language PK4::language(void) const { return Language(data[0x17]); }
     void PK4::language(Language v) { data[0x17] = u8(v); }
 
-    u8 PK4::ev(Stat ev) const { return data[0x18 + u8(ev)]; }
-    void PK4::ev(Stat ev, u8 v) { data[0x18 + u8(ev)] = v; }
+    u16 PK4::ev(Stat ev) const { return data[0x18 + u8(ev)]; }
+    void PK4::ev(Stat ev, u16 v) { data[0x18 + u8(ev)] = v; }
 
     u8 PK4::contest(u8 contest) const { return data[0x1E + contest]; }
     void PK4::contest(u8 contest, u8 v) { data[0x1E + contest] = v; }
