+++ source/pkx/PK6.cpp
@@ -317,8 +317,8 @@ namespace pksm
     u16 PK6::alternativeForm(void) const { return data[0x1D] >> 3; }
     void PK6::alternativeForm(u16 v) { data[0x1D] = (data[0x1D] & 0x07) | (v << 3); }
 
-    u8 PK6::ev(Stat ev) const { return data[0x1E + u8(ev)]; }
-    void PK6::ev(Stat ev, u8 v) { data[0x1E + u8(ev)] = v; }
+    u16 PK6::ev(Stat ev) const { return data[0x1E + u8(ev)]; }
+    void PK6::ev(Stat ev, u16 v) { data[0x1E + u8(ev)] = v; }
 
     u8 PK6::contest(u8 contest) const { return data[0x24 + contest]; }
     void PK6::contest(u8 contest, u8 v) { data[0x24 + contest] = v; }
